third_party/WebKit/LayoutTests/external/wpt/hr-time/timeOrigin_with_timeout.html

<!doctype html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
async_test(function(t) {
  this.step_timeout(function() {
    const workerScript = 'postMessage({timeOrigin: performance.timeOrigin})';
    const blob = new Blob([workerScript]);
    const workerCreation = performance.now();
    const worker = new Worker(URL.createObjectURL(blob));
    worker.addEventListener('message', t.step_func_done(function(event) {
      const windowOrigin = performance.timeOrigin;
      const workerOrigin = event.data.timeOrigin;
      assert_less_than(workerOrigin - windowOrigin - workerCreation, 100, 'Difference between worker and window timeOrigins should roughly be equal to performance.now() at time of creation.');
      assert_greater_than_equal(workerOrigin - windowOrigin, 200, 'We waited 200ms to spawn the second worker, so the timeOrigin should be much greater.');
    }));
  }, 200);
}, 'Window and worker timeOrigins differ when the worker is created after a timeout.');
</script>
</body>
</html>
